Returned Lost Nexus to Map Hell.
Slight 1% shrink to Thunder Piston to try and fix the jump issues with some characters. Hopefully playable.

Reverted FIsh's name changes to Neo Test Track.
Fish had changed Neo Test Track to 1 lap. Which is kind of silly, should have just hidden it. Instead I propose a compromise and set it to 7 laps instead of the previous 15 or 1 laps.